renamings and small fixes
