+++ crates/control/src/behavior/support.rs
@@ -5,7 +5,7 @@ use geometry::look_at::LookAt;
 use nalgebra::{point, Isometry2, UnitComplex, Vector2};
 use types::{
     field_dimensions::FieldDimensions,
-    filtered_game_states::FilteredGameState,
+    filtered_game_state::FilteredGameState,
     motion_command::MotionCommand,
     path_obstacles::PathObstacle,
     support_foot::Side,
@@ -63,7 +63,7 @@ fn support_pose(
         Some(FilteredGameState::Ready { .. })
         | Some(FilteredGameState::Playing {
             ball_is_free: false,
-            kick_off: _,
+            ..
         }) => supporting_position.x.clamp(
             minimum_x.min(maximum_x_in_ready_and_when_ball_is_not_free),
             minimum_x.max(maximum_x_in_ready_and_when_ball_is_not_free),
