+++ crates/types/src/filtered_game_state.rs
