+++ crates/types/src/lib.rs
@@ -18,7 +18,7 @@ pub mod field_color;
 pub mod field_dimensions;
 pub mod field_marks;
 pub mod filtered_game_controller_state;
-pub mod filtered_game_states;
+pub mod filtered_game_state;
 pub mod filtered_segments;
 pub mod filtered_whistle;
 pub mod game_controller_state;
