+++ etc/parameters/head.P0000074A06S9A900031.json
@@ -2,16 +2,12 @@
   "camera_matrix_parameters": {
     "vision_bottom": {
       "extrinsic_rotations": [
-        0.2427273988723755,
-        -4.1517486572265625,
-        1.1114037036895752
+        0.2427273988723755, -4.1517486572265625, 1.1114037036895752
       ]
     },
     "vision_top": {
       "extrinsic_rotations": [
-        0.14052987098693848,
-        -5.275916576385498,
-        1.6634505987167358
+        0.14052987098693848, -5.275916576385498, 1.6634505987167358
       ]
     }
   }
