+++ etc/parameters/nao_location
@@ -1,1 +1,1 @@
-leipzig
+smd
